WIP: improving code quality
- removing redundant code
